backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
email-validator==2.1.0
bcrypt==4.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
psutil==5.9.6
psycopg2-binary==2.9.9

# Linting and Type Checking
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1

# For Python code execution
subprocess32==3.5.4; sys_platform != "win32"

# For secure code execution with Kubernetes
kubernetes==29.0.0

# Additional utilities
pandas==2.1.3
numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.2